重构项目结构的类型定义以及导出
