+++ src/index.ts
@@ -1,8 +1,7 @@
-export * as constants from "./constants";
+export * from "./constants";
 export { default as i18n } from "./i18n";
-export * from "./locales/typing";
-export * from "./models/typing";
 export * from "./tools/date";
 export * from "./tools/i18n";
 export * from "./tools/math";
+export * from "./typings";
 export * as ziwei from "./ziwei";
